+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Namespaces `WUP...` is refactored from `WUPPopup` to `WUP.Popup`. So re-export enums `ShowCases, HideCases, Animations` without prefix now
 - helper localeInfo moved to `web-ui-pack/objects/localeInfo`
+- helper scrollCarousel refactored to [WUPScrolled](src/helpers/scrolled.ts) class
 
 **Fixes**: